% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrcity-file-operations.R
\name{convert_session_to_date}
\alias{convert_session_to_date}
\title{Converts a session name to postix date}
\usage{
convert_session_to_date(string)
}
\arguments{
\item{string}{}
}
\value{

}
\description{
Converts a session name to postix date
}
