% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrcity-quest-navigationextended.R
\name{get_navigation_target}
\alias{get_navigation_target}
\title{Returns english code for a navigation target}
\usage{
get_navigation_target(task)
}
\arguments{
\item{task}{}
}
\value{

}
\description{
Returns english code for a navigation target
}
