% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrcity-quest-navigationextended.R
\name{get_navigation_target_position}
\alias{get_navigation_target_position}
\title{Returns x, y position of a target site}
\usage{
get_navigation_target_position(task)
}
\arguments{
\item{task}{}
}
\value{

}
\description{
Returns x, y position of a target site
}
