% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrcity-getters.R
\name{get_started_quests}
\alias{get_started_quests}
\title{Returns list with quest and their ids which have been started}
\usage{
get_started_quests(schedule)
}
\arguments{
\item{schedule}{}
}
\value{

}
\description{
Returns list with quest and their ids which have been started
}
