% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrcity-getters.R
\name{get_task_start_position}
\alias{get_task_start_position}
\title{Returns task start position}
\usage{
get_task_start_position(task)
}
\arguments{
\item{task}{}
}
\value{
numeric(2) or NULL
}
\description{
Returns task start position
}
