% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrcity-getters.R
\name{get_task_start_time}
\alias{get_task_start_time}
\title{Returns task start time}
\usage{
get_task_start_time(task)
}
\arguments{
\item{task}{}
}
\value{
numeric(1) with the start time or NULL if not found
}
\description{
Returns task start time
}
