% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrcity-getters.R
\name{get_tasks_by_name}
\alias{get_tasks_by_name}
\title{Returns tasks of given name from the expeiremnt. If you know the task id, use `get_task`}
\usage{
get_tasks_by_name(experiment, taskname)
}
\arguments{
\item{experiment}{expeirment object}

\item{taskname}{name of the task to extract. Task names can be accessed through
the QUEST_NAMES exported list (e.g. `QUEST_NAMES$citynavigation`)}
}
\value{
list of tasks
}
\description{
Returns tasks of given name from the expeiremnt. If you know the task id, use `get_task`
}
