% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrcity-loading.R
\name{load_experiment}
\alias{load_experiment}
\title{Laods the experiemnt data from given folder into a vrcity class}
\usage{
load_experiment(folder)
}
\arguments{
\item{folder}{folder with all the subfolders}
}
\value{
list of crcity class
}
\description{
Laods the experiemnt data from given folder into a vrcity class
}
