% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrcity-loading.R
\name{load_schedule_log}
\alias{load_schedule_log}
\title{Loads the main schedule log}
\usage{
load_schedule_log(folder)
}
\arguments{
\item{folder}{sesion folder with the log and tasks present}
}
\value{
returns shcdule log list
}
\description{
Loads the main schedule log
}
