% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrcity-loading.R
\name{load_task}
\alias{load_task}
\title{Loads a single task in given folder}
\usage{
load_task(folder)
}
\arguments{
\item{folder}{path to the task data}
}
\value{
list with a vretcity task class
}
\description{
Loads a single task in given folder
}
